+++ Platform-io-source/src/tw_controls/control_Button.h
@@ -1,10 +1,10 @@
-#include "tw_controls/tw_control.h"
 #include "settings/settingsOption.h"
+#include "tw_controls/tw_control.h"
 
 class ControlButton : public tw_control
 {
 	public:
-	    void draw(uint canvasid);
+		void draw(uint8_t canvasid);
 		bool drag(int16_t drag_x, int16_t drag_y);
 		void drag_end();
 		bool click(uint16_t click_pos_x, uint16_t click_pos_y);
@@ -13,5 +13,4 @@ class ControlButton : public tw_control
 
 	private:
 		String version = "1.0";
-
 };
