+++ Platform-io-source/src/tw_controls/control_Toggle.h
@@ -1,10 +1,10 @@
-#include "tw_controls/tw_control.h"
 #include "settings/settingsOption.h"
+#include "tw_controls/tw_control.h"
 
 class ControlToggle : public tw_control
 {
 	public:
-	    void draw(uint canvasid);
+		void draw(uint8_t canvasid);
 		bool drag(int16_t drag_x, int16_t drag_y);
 		void drag_end();
 		bool click(uint16_t click_pos_x, uint16_t click_pos_y);
@@ -12,10 +12,7 @@ class ControlToggle : public tw_control
 		bool click_long(uint16_t click_pos_x, uint16_t click_pos_y);
 		void set_data(SettingsOptionBool *sett);
 
-		
-
 	private:
 		SettingsOptionBool *setting_option = nullptr;
 		String version = "1.0";
-
 };
