+++ Platform-io-source/src/tw_controls/tw_control.h
@@ -2,7 +2,6 @@
 
 #include "display.h"
 
-
 extern TFT_eSPI tft;
 extern TFT_eSprite canvas[];
 
@@ -10,8 +9,7 @@ class tw_face;
 
 class tw_control
 {
-    public:
-
+	public:
 		using CallbackFunction = void (*)();
 
 		void create(String _name, String _option1, String _option2, uint _pos_x, uint _pos_y, uint _width, uint _height);
@@ -20,7 +18,7 @@ class tw_control
 		void set_callback(CallbackFunction callback);
 		void set_values(String _option1, String _option2);
 		void set_parent(tw_face *_parent);
-		bool bounds_check(uint click_pos_x, uint click_pos_y);
+		bool bounds_check(int16_t click_pos_x, int16_t click_pos_y);
 		uint16_t get_height();
 		uint16_t get_height_stacked();
 
@@ -31,23 +29,21 @@ class tw_control
 		void draw_scroll(uint canvasid, uint16_t offset_x, uint16_t offset_y);
 
 		void set_scrollable(bool scroll_x, bool scroll_y);
-		bool can_drag(uint click_pos_x, uint click_pos_y);
+		bool can_drag(int16_t click_pos_x, int16_t click_pos_y);
 
 		// Virtual functions
-        virtual void draw(uint canvasid) = 0;
+		virtual void draw(uint8_t canvasid) = 0;
 		virtual bool drag(int16_t drag_x, int16_t drag_y) = 0;
 		virtual void drag_end() = 0;
-		virtual bool click(uint16_t click_pos_x, uint16_t click_pos_y) {return false;}
-		virtual bool click_double(uint16_t click_pos_x, uint16_t click_pos_y) {return false;}
-		virtual bool click_long(uint16_t click_pos_x, uint16_t click_pos_y) {return false;}
+		virtual bool click(uint16_t click_pos_x, uint16_t click_pos_y) { return false; }
+		virtual bool click_double(uint16_t click_pos_x, uint16_t click_pos_y) { return false; }
+		virtual bool click_long(uint16_t click_pos_x, uint16_t click_pos_y) { return false; }
 		String name = "";
 
-		
-
-    protected:
+	protected:
 		tw_face *parent = nullptr;
 
-        // TFT_eSprite my_sprite = TFT_eSprite(&tft);
+		// TFT_eSprite my_sprite = TFT_eSprite(&tft);
 
 		CallbackFunction callbackFunction;
 
@@ -55,7 +51,7 @@ class tw_control
 		String option1 = "";
 		String option2 = "";
 
-		uint canvasid = 0;
+		uint8_t canvasid = 0;
 		uint width = 0;
 		uint height = 0;
 		uint pos_x = 0;
@@ -63,7 +59,7 @@ class tw_control
 		uint padding_x = 20;
 		uint padding_y = 20;
 
-        bool requires_redraw = true;
+		bool requires_redraw = true;
 
 		bool can_scroll_x = false;
 		bool can_scroll_y = false;
